server/requirements.txt
# --- Core dependencies ---
fastapi==0.120.0
uvicorn[standard]==0.38.0

# --- Database ---
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# --- Migrations ---
alembic==1.17.0

# --- Env / Settings management ---
python-dotenv==1.1.1

# --- Optional utilities ---
pydantic==2.12.3
pydantic[email]==2.12.3
pwdlib[argon2]
python-jose[cryptography]
python-multipart
PyJWT
python-slugify